moved select menus in different component folder
