+++ components/profile.jsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from 'react'
 import { useSupabaseClient, useUser } from '@supabase/auth-helpers-react'
 import { UserCircleIcon } from '@heroicons/react/24/solid'
 import RadioGroup from './radioGroup'
-import BackgroundSelectMenu from './selectMenu'
+import BackgroundSelectMenu from './selectMenus/selectMenu'
 
 export default function Profile() {
   const supabase = useSupabaseClient()
@@ -99,8 +99,7 @@ export default function Profile() {
         <div className='flex justify-center px-4 md:px-0 bg-white'>
           <form
             onSubmit={updateProfile}
-            className='max-w-4xl flex flex-col justify-center'
-          >
+            className='max-w-4xl flex flex-col justify-center'>
             <div className='space-y-12'>
               <div className='border-b border-gray-900/10 pb-12'>
                 <h2 className='text-xl font-semibold leading-7 text-gray-900'>
@@ -115,8 +114,7 @@ export default function Profile() {
                   <div className='col-span-full'>
                     <label
                       htmlFor='photo'
-                      className='block text-sm font-medium leading-6 text-gray-900'
-                    >
+                      className='block text-sm font-medium leading-6 text-gray-900'>
                       Profile Picture
                     </label>
                     <div className='mt-2 flex items-center gap-x-3'>
@@ -126,8 +124,7 @@ export default function Profile() {
                       />
                       <button
                         type='button'
-                        className='rounded-md bg-white px-2.5 py-1.5 pl-2 text-sm font-semibold text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 hover:bg-gray-50'
-                      >
+                        className='rounded-md bg-white px-2.5 py-1.5 pl-2 text-sm font-semibold text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 hover:bg-gray-50'>
                         Change
                       </button>
                     </div>
@@ -135,8 +132,7 @@ export default function Profile() {
                   <div className='sm:col-span-4'>
                     <label
                       htmlFor='username'
-                      className='block text-sm font-medium leading-6 text-gray-900'
-                    >
+                      className='block text-sm font-medium leading-6 text-gray-900'>
                       Username
                     </label>
                     <div className='mt-2'>
@@ -158,8 +154,7 @@ export default function Profile() {
                   <div className='col-span-full'>
                     <label
                       htmlFor='about'
-                      className='block text-sm font-medium leading-6 text-gray-900'
-                    >
+                      className='block text-sm font-medium leading-6 text-gray-900'>
                       About
                     </label>
                     <p className='mt-3 text-sm leading-6 text-gray-600'>
@@ -199,8 +194,7 @@ export default function Profile() {
                   <div className='sm:col-span-3'>
                     <label
                       htmlFor='first-name'
-                      className='block text-sm font-medium leading-6 text-gray-900'
-                    >
+                      className='block text-sm font-medium leading-6 text-gray-900'>
                       First name
                     </label>
                     <div className='mt-2'>
@@ -219,8 +213,7 @@ export default function Profile() {
                   <div className='sm:col-span-3'>
                     <label
                       htmlFor='last-name'
-                      className='block text-sm font-medium leading-6 text-gray-900'
-                    >
+                      className='block text-sm font-medium leading-6 text-gray-900'>
                       Last name
                     </label>
                     <div className='mt-2'>
@@ -239,8 +232,7 @@ export default function Profile() {
                   <div className='sm:col-span-4'>
                     <label
                       htmlFor='email'
-                      className='block text-sm font-medium leading-6 text-gray-900'
-                    >
+                      className='block text-sm font-medium leading-6 text-gray-900'>
                       Email address
                     </label>
                     <div className='mt-2'>
@@ -259,8 +251,7 @@ export default function Profile() {
                   <div className='sm:col-span-3'>
                     <label
                       htmlFor='country'
-                      className='block text-sm font-medium leading-6 text-gray-900'
-                    >
+                      className='block text-sm font-medium leading-6 text-gray-900'>
                       Country
                     </label>
                     <div className='mt-2'>
@@ -268,8 +259,7 @@ export default function Profile() {
                         id='country'
                         name='country'
                         autoComplete='country-name'
-                        className='block w-full rounded-md border-0 py-1.5 pl-2 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 focus:ring-2 focus:ring-inset focus:ring-blue-600 sm:max-w-xs sm:text-sm sm:leading-6'
-                      >
+                        className='block w-full rounded-md border-0 py-1.5 pl-2 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 focus:ring-2 focus:ring-inset focus:ring-blue-600 sm:max-w-xs sm:text-sm sm:leading-6'>
                         <option>United States</option>
                         <option>Canada</option>
                         <option>Mexico</option>
@@ -280,8 +270,7 @@ export default function Profile() {
                   <div className='col-span-full'>
                     <label
                       htmlFor='street-address'
-                      className='block text-sm font-medium leading-6 text-gray-900'
-                    >
+                      className='block text-sm font-medium leading-6 text-gray-900'>
                       Street address
                     </label>
                     <div className='mt-2'>
@@ -300,8 +289,7 @@ export default function Profile() {
                   <div className='sm:col-span-2 sm:col-start-1'>
                     <label
                       htmlFor='city'
-                      className='block text-sm font-medium leading-6 text-gray-900'
-                    >
+                      className='block text-sm font-medium leading-6 text-gray-900'>
                       City
                     </label>
                     <div className='mt-2'>
@@ -320,8 +308,7 @@ export default function Profile() {
                   <div className='sm:col-span-2'>
                     <label
                       htmlFor='region'
-                      className='block text-sm font-medium leading-6 text-gray-900'
-                    >
+                      className='block text-sm font-medium leading-6 text-gray-900'>
                       State / Province
                     </label>
                     <div className='mt-2'>
@@ -340,8 +327,7 @@ export default function Profile() {
                   <div className='sm:col-span-2'>
                     <label
                       htmlFor='postal-code'
-                      className='block text-sm font-medium leading-6 text-gray-900'
-                    >
+                      className='block text-sm font-medium leading-6 text-gray-900'>
                       ZIP / Postal code
                     </label>
                     <div className='mt-2'>
@@ -363,14 +349,12 @@ export default function Profile() {
             <div className='mt-6 flex items-center justify-end gap-x-6'>
               <button
                 type='button'
-                className='text-sm font-semibold leading-6 text-gray-900'
-              >
+                className='text-sm font-semibold leading-6 text-gray-900'>
                 Cancel
               </button>
               <button
                 type='submit'
-                className='rounded-md bg-blue-600 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-blue-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-blue-600'
-              >
+                className='rounded-md bg-blue-600 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-blue-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-blue-600'>
                 Save
               </button>
             </div>
